insert multi item

+++ internal/model/web/item_request.go
@@ -1,7 +1,7 @@
 package web
 
 type ItemRequest []struct {
-	Qty    int     `json:"qty"`
+	Qty    float64 `json:"qty"`
 	Price  float64 `json:"price"`
 	Amount float64 `json:"amount"`
 }
